requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
Django==2.2.5
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.10.3
fabric==2.5.0
invoke==1.3.0
paramiko==2.6.0
Pillow==6.1.0
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.2
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
wincertstore==0.2
